5KMZI

+++ KMZI/lab5/task.cpp
@@ -2,11 +2,11 @@
 #include <cmath>
 using namespace std;
 
-bool is_prime(int n) {
+bool is_prime(unsigned long long n) {
     if (n < 2) {
         return false;
     }
-    for (int i = 2; i <= sqrt(n); i++) {
+    for (unsigned long long i = 2; i <= sqrt(n); i++) {
         if (n % i == 0) {
             return false;
         }
@@ -14,9 +14,9 @@ bool is_prime(int n) {
     return true;
 }
 
-int generate_prime(int k) {
-    int n = pow(2, k) - 1;
-    int i = 2;
+unsigned long long generate_prime(unsigned long long k) {
+    unsigned long long n = pow(2, k) - 1;
+    unsigned long long i = 2;
     while (!is_prime(n)) {
         n = pow(2, k) - 1;
         n = n * i;
@@ -25,10 +25,10 @@ int generate_prime(int k) {
     return n;
 }
 
-int main() {
-    int k;
+unsigned long long main() {
+    unsigned long long k;
     cout << "Enter k: ";
     cin >> k;
-    int prime = generate_prime(k);
+    unsigned long long prime = generate_prime(k);
     cout << "Prime number: " << prime << endl;
  }
